Add files via upload
Program that converts an xlsx file to and xml file. The converted file is sent to whatever address is established; in this case it is Wesco. Once the file has been sent the files are deleted. 

When preparing an xlsx file, save it in the same directory as this file, or change the location in the DeleteFile module.